Move models into submodules and update imports.
