+++ open_rarity/models/__init__.py
@@ -1,6 +1,6 @@
 from .collection import Collection
 from .token import Token
-from .token_identifier import EVMContractTokenIdentifier
-from .token_metadata import StringAttribute, TokenMetadata
-from .token_rarity import TokenRarity
-from .token_standard import TokenStandard
+from .token.identifier import EVMContractTokenIdentifier
+from .token.metadata import StringAttribute, TokenMetadata
+from .token.rarity import TokenRarity
+from .token.standard import TokenStandard
