+++ open_rarity/models/collection/__init__.py
@@ -1,0 +1,1 @@
+from .collection import Collection, CollectionAttribute
