+++ open_rarity/models/token/identifier.py
