+++ open_rarity/models/token/metadata.py
