+++ open_rarity/models/token/rarity.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
-from open_rarity.models.token import Token
-from open_rarity.models.token_ranking_features import TokenRankingFeatures
+from .ranking_features import TokenRankingFeatures
+from .token import Token
 
 
 @dataclass
