+++ open_rarity/rarity_ranker.py
@@ -1,9 +1,9 @@
 import math
 
 from open_rarity.models.collection import Collection
-from open_rarity.models.token_rarity import TokenRarity
+from open_rarity.models.token.rarity import TokenRarity
+from open_rarity.scoring.feature_extractor import TokenFeatureExtractor
 from open_rarity.scoring.scorer import Scorer
-from open_rarity.scoring.token_feature_extractor import TokenFeatureExtractor
 
 
 class RarityRanker:
@@ -103,7 +103,7 @@ class RarityRanker:
         sorted_token_rarities: list[TokenRarity] = sorted(
             token_rarities,
             key=lambda k: (
-                k.token_features.unique_attribute_count,
+                k.token.features.unique_attribute_count,
                 k.score,
             ),
             reverse=True,
