+++ open_rarity/resolver/opensea_api_helpers.py
@@ -7,14 +7,14 @@ from requests.models import HTTPError
 
 from open_rarity.models.collection import Collection
 from open_rarity.models.token import Token
-from open_rarity.models.token_identifier import EVMContractTokenIdentifier
-from open_rarity.models.token_metadata import (
+from open_rarity.models.token.identifier import EVMContractTokenIdentifier
+from open_rarity.models.token.metadata import (
     DateAttribute,
     NumericAttribute,
     StringAttribute,
     TokenMetadata,
 )
-from open_rarity.models.token_standard import TokenStandard
+from open_rarity.models.token.standard import TokenStandard
 from open_rarity.resolver.models.collection_with_metadata import CollectionWithMetadata
 
 logger = logging.getLogger("open_rarity_logger")
