+++ open_rarity/resolver/rarity_providers/external_rarity_provider.py
@@ -4,7 +4,7 @@ from collections import defaultdict
 
 import requests
 
-from open_rarity.models.token_identifier import EVMContractTokenIdentifier
+from open_rarity.models.token.identifier import EVMContractTokenIdentifier
 from open_rarity.resolver.models.collection_with_metadata import CollectionWithMetadata
 from open_rarity.resolver.models.token_with_rarity_data import (
     EXTERNAL_RANK_PROVIDERS,
