+++ open_rarity/scoring/handlers/arithmetic_mean_scoring_handler.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from open_rarity.models.collection import Collection, CollectionAttribute
 from open_rarity.models.token import Token
-from open_rarity.models.token_metadata import AttributeName
+from open_rarity.models.token.metadata import AttributeName
 from open_rarity.scoring.utils import get_token_attributes_scores_and_weights
 
 logger = logging.getLogger("open_rarity_logger")
