+++ open_rarity/scoring/scorer.py
@@ -1,6 +1,6 @@
 from open_rarity.models.collection import Collection
 from open_rarity.models.token import Token
-from open_rarity.models.token_standard import TokenStandard
+from open_rarity.models.token.standard import TokenStandard
 from open_rarity.scoring.handlers.information_content_scoring_handler import (
     InformationContentScoringHandler,
 )
