+++ open_rarity/scoring/utils.py
@@ -1,6 +1,6 @@
 from open_rarity.models.collection import Collection, CollectionAttribute
 from open_rarity.models.token import Token
-from open_rarity.models.token_metadata import AttributeName
+from open_rarity.models.token.metadata import AttributeName
 
 
 def get_token_attributes_scores_and_weights(
