+++ tests/helpers.py
@@ -2,17 +2,17 @@ from random import shuffle
 
 from open_rarity.models.collection import Collection
 from open_rarity.models.token import Token
-from open_rarity.models.token_identifier import (
+from open_rarity.models.token.identifier import (
     EVMContractTokenIdentifier,
     SolanaMintAddressTokenIdentifier,
 )
-from open_rarity.models.token_metadata import (
+from open_rarity.models.token.metadata import (
     AttributeName,
     NumericAttribute,
     StringAttribute,
     TokenMetadata,
 )
-from open_rarity.models.token_standard import TokenStandard
+from open_rarity.models.token.standard import TokenStandard
 
 
 def create_evm_token(
