+++ tests/models/test_collection.py
@@ -1,7 +1,7 @@
 from open_rarity.models.collection import Collection, CollectionAttribute
 from open_rarity.models.token import Token
-from open_rarity.models.token_metadata import StringAttribute, TokenMetadata
-from open_rarity.models.token_standard import TokenStandard
+from open_rarity.models.token.metadata import StringAttribute, TokenMetadata
+from open_rarity.models.token.standard import TokenStandard
 from tests.helpers import (
     create_evm_token,
     create_numeric_evm_token,
