+++ tests/models/test_token.py
@@ -1,14 +1,14 @@
 from open_rarity.models.token import Token
-from open_rarity.models.token_identifier import (
+from open_rarity.models.token.identifier import (
     EVMContractTokenIdentifier,
     SolanaMintAddressTokenIdentifier,
 )
-from open_rarity.models.token_metadata import (
+from open_rarity.models.token.metadata import (
     NumericAttribute,
     StringAttribute,
     TokenMetadata,
 )
-from open_rarity.models.token_standard import TokenStandard
+from open_rarity.models.token.standard import TokenStandard
 from tests.helpers import create_evm_token
